[pr-2] Move LegacyStaticEmbeddingModalProps to the plugins contract
Module `embedding-iframe-sdk-setup` will be moved to app tier, so we move
`LegacyStaticEmbeddingModalProps` to the oss plugin file to avoid an import
going up to app
