+++ frontend/src/metabase/embedding/embedding-iframe-sdk-setup/types.ts
@@ -14,12 +14,7 @@ import type {
 
 export type { SdkIframeEmbedSetupTheme } from "metabase-types/api";
 
-export type SdkIframeEmbedSetupExperience =
-  | "dashboard"
-  | "chart"
-  | "exploration"
-  | "browser"
-  | "metabot";
+export type { SdkIframeEmbedSetupExperience } from "metabase/plugins/oss/embedding-iframe-sdk-setup";
 
 export type SdkIframeEmbedSetupStep =
   | "select-embed-experience"
