+++ frontend/src/metabase/new/components/NewModals/NewModals.tsx
@@ -10,14 +10,14 @@ import CreateCollectionModal, {
 import { useInitialCollectionId } from "metabase/common/collections/hooks";
 import { UpgradeModal } from "metabase/common/components/upsells/components/UpgradeModal";
 import { STATIC_LEGACY_EMBEDDING_TYPE } from "metabase/embedding/constants";
-import {
-  LegacyStaticEmbeddingModal,
-  type LegacyStaticEmbeddingModalProps,
-} from "metabase/embedding/embedding-iframe-sdk-setup/components/LegacyStaticEmbeddingModal";
+import { LegacyStaticEmbeddingModal } from "metabase/embedding/embedding-iframe-sdk-setup/components/LegacyStaticEmbeddingModal";
 import { SdkIframeEmbedSetupModal } from "metabase/embedding/embedding-iframe-sdk-setup/components/SdkIframeEmbedSetupModal";
 import { PaletteShortcutsModal } from "metabase/palette/components/PaletteShortcutsModal/PaletteShortcutsModal";
 import { useRegisterShortcut } from "metabase/palette/hooks/useRegisterShortcut";
-import type { SdkIframeEmbedSetupModalProps } from "metabase/plugins";
+import type {
+  LegacyStaticEmbeddingModalProps,
+  SdkIframeEmbedSetupModalProps,
+} from "metabase/plugins";
 import { useDispatch, useSelector } from "metabase/redux";
 import { closeModal, setOpenModal } from "metabase/redux/ui";
 import type { WithRouterProps } from "metabase/router";
