+++ frontend/src/metabase/plugins/index.ts
@@ -78,6 +78,8 @@ export {
   PLUGIN_EMBEDDING_IFRAME_SDK_SETUP,
   type SdkIframeEmbedSetupModalProps,
   type SdkIframeEmbedSetupModalInitialState,
+  type SdkIframeEmbedSetupExperience,
+  type LegacyStaticEmbeddingModalProps,
 } from "./oss/embedding-iframe-sdk-setup";
 export { PLUGIN_EMBEDDING_SDK } from "./oss/embedding-sdk";
 export {
