+++ frontend/src/metabase/plugins/oss/embedding-iframe-sdk-setup.ts
@@ -1,3 +1,5 @@
+import type { DashboardId } from "metabase-types/api";
+
 export type SdkIframeEmbedSetupModalProps = {
   opened: boolean;
   onClose: () => void;
@@ -9,6 +11,20 @@ export type SdkIframeEmbedSetupModalInitialState = {
   resourceId?: string | number | null;
   isGuest?: boolean;
   useExistingUserSession?: boolean;
+};
+
+export type SdkIframeEmbedSetupExperience =
+  | "dashboard"
+  | "chart"
+  | "exploration"
+  | "browser"
+  | "metabot";
+
+export type LegacyStaticEmbeddingModalProps = {
+  experience: SdkIframeEmbedSetupExperience;
+  dashboardId?: DashboardId | null;
+  questionId?: string | number | null;
+  parentInitialState: SdkIframeEmbedSetupModalInitialState;
 };
 
 const getDefaultPluginEmbeddingIframeSdkSetup = () => ({
